feat(PlaybackControlUI): 음악 재생 바인드 서비스 UI 컴포저블 분리
- `PlaybackControlUI` 컴포저블 함수를 `PlaybackBoundServiceActivity`에서 별도 파일로 분리
